requirements.txt
numpy==1.15.1
scipy==1.1.0
torch==1.2.0
torchvision==0.4.0
avalanche-lib==0.4.0
typing-extensions==4.4.0
psutil
gputil
scikit-learn
matplotlib
numpy
pytorchcv
wandb
tensorboard>=1.15
tqdm
torch
torchvision
torchmetrics
gdown
dill
packaging
quadprog